Fix incompatibility with old versions of the su binary
It may seem as though this will hurt performance, but in my testing the
app is actually faster this way. There is a very minor decrease in
performance when using a 2.x binary, but that's almost exclusively in
the binary itself. Perhaps this is how Superuser should have worked all
along.