Fix FCs happening when users press back while the binary updater is running
There are two different kinds of canceling happening here, if back is
pressed while a file is being downloaded, the updater stops
immediately. If the user presses back once the updater reaches the
"critical phase", when the filesystem operations are actually
happening, the updater will finish if possible, just behind the scenes.